tests/wpt/web-platform-tests/client-hints/accept_ch_lifetime_subresource.tentative.https.html

<html>
<title>Accept-CH-Lifetime test with subresource</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
Apart from this webpage, the test opens another html web page. One test is run
in this web page, and another in the second web page.
-->

<script>

// This test fetches resources/accept_ch_lifetime.html as a subresource. The
// response headers to that webpage contain Accept-CH and Accept-CH-Lifetime
// headers.

// Fetching that webpage as a subresource should NOT cause the user-agent to
// persist origin preferences for the client hints specified in Accept-CH
// header.

// Next, to verify if the origin preferences were not persisted by the user
// agent, this test fetches resources/do_not_expect_client_hints_headers.html
// in a new window. Fetching of
// resources/do_not_expect_client_hints_headers.html verifies that the user
// agent does not send the client hints in the request headers.

// Test is marked as tentative until https://github.com/whatwg/fetch/issues/726
// is resolved.

// First, verify the initial state to make sure that the browser does not have
// client hints preferences cached from a previous run of the test.
promise_test(t => {
  return fetch("echo_client_hints_received.py").then(r => {
    assert_equals(r.status, 200)
    // Verify that the browser did not include client hints in the request
    // headers when fetching echo_client_hints_received.py.
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
  });
}, "Precondition: Test that the browser does not have client hints preferences cached");

promise_test(t => {
  // Fetching this web page as a subresource should NOT cause user-agent to
  // persist client hint preferences for the origin.
  return fetch("resources/accept_ch_lifetime.html").then(r => {
    assert_equals(r.status, 200)
    // Verify that the browser did not include client hints in the request
    // headers.
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
    // Open a new window. Verify that the user agent does not attach the client
    // hints.
    var win = window.open("resources/do_not_expect_client_hints_headers.html");
    assert_not_equals(win, null, "Popup windows not allowed?");
  });
}, "Test receiving Accept-CH-Lifetime header");

async_test(t => {
  window.addEventListener('message', function(e) {
    if(!e.source.location.pathname.includes("do_not_expect_client_hints_headers.html")) {
      return;
    }
    if(typeof e.data != "string")
      return;
    assert_equals(e.data, "PASS");
    t.done();
  })
}, "Loading of resources/do_not_expect_client_hints_headers.html did not finish.");

</script>
</body>
</html>
